Adding similarity score, pending testing

+++ src/main/java/com/tesis/resilientest/resilient/ElementScore.java
@@ -1,8 +1,8 @@
 package com.tesis.resilientest.resilient;
 
 import org.openqa.selenium.WebElement;
-public record ElementScore(WebElement element, int score) {
-    public ElementScore(WebElement element, int score) {
+public record ElementScore(WebElement element, double score) {
+    public ElementScore(WebElement element, double score) {
         if (score < 0 || score > 100) {
             throw new IllegalArgumentException("Score must be between 0 and 100");
         }
@@ -10,7 +10,7 @@ public record ElementScore(WebElement element, int score) {
         this.score = score;
     }
 
-    public static ElementScore of(WebElement element, int score) {
+    public static ElementScore of(WebElement element, double score) {
         return new ElementScore(element, score);
     }
 }
